add both parser flags
1. It is possible to use --insecure-http-parser in the CLI or NODE_OPTIONS to configure this per-process.
2. Pass insecureHTTPParser to http.createServer() to configure this on a per-server basis.
3. Pass insecureHTTPParser to http.request() to configure this on a per-stream basis.
4. Use --http-parser=legacy CLI options (only available on v12).

For v13: 1st option is available since v13.4.0 while 2nd and 3rd since v13.8.0.

For v12: the first 3 options are available since v12.15.0 and the 4th since the start of v12.

For v10: the first 3 options are available since v10.19.0.

The 4th option has been available since v11.4.0 and is therefore available in v12 but was removed since v13.0.0.